Backend/ocr/requirement.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
numpy==1.24.3
opencv-python==4.10.0.84
pyocr==0.8.5
pillow==10.4.0
Werkzeug==3.0.6
matplotlib==3.7.5
